Big overhaul! No more individual modules per service. The AnyDelete can delete most services.
